requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-decouple==3.8

# API Integration
requests==2.31.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Documentation
drf-yasg==1.21.7
